+++ tests/opentelemetry-docker-tests/tests/sqlalchemy_tests/test_instrument.py
@@ -60,8 +60,9 @@ class SQLAlchemyInstrumentTestCase(TestBase):
 
         traces = self.memory_exporter.get_finished_spans()
         # trace composition
-        self.assertEqual(len(traces), 1)
-        span = traces[0]
+        # one span for the connection and one for the query
+        self.assertEqual(len(traces), 2)
+        span = traces[1]
         # check subset of span fields
         self.assertEqual(span.name, "SELECT opentelemetry-tests")
         self.assertIs(span.status.status_code, trace.StatusCode.UNSET)